+++ @udir-design/react/src/typography/index.tsx
@@ -1,4 +1,4 @@
 export * from './heading/Heading';
 export * from './label/Label';
 export * from './paragraph/Paragraph';
-export * from './validation-message/ValidationMessage';
+export * from './validationMessage/ValidationMessage';
