+++ @udir-design/react/src/typography/validationMessage/ValidationMessage.stories.tsx
@@ -2,7 +2,6 @@ import type { Meta, StoryObj } from '@storybook/react';
 import { ValidationMessage } from './ValidationMessage';
 
 const meta: Meta<typeof ValidationMessage> = {
-  title: 'ValidationMessage',
   component: ValidationMessage,
 };
